# swagger.json
{
  "info": {
    "title": "SWAGGER API",
    "version": "1.0.0",
    "description": "SWAGGER API  with autogenerated swagger doc"
  },
  "components": {
    "securitySchemes": {
      "type": "http",
      "scheme": "basic"
    }
  },
  "basePath": "/api/",
  "swagger": "2.0",
  "paths": {
    "/search": {
      "get": {
        "tags": [
          "test"
        ],
        "description": "poi search",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "description": "KYC Info key code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "An array of users"
          }
        }
      }
    },
    "/res": {
      "get": {
        "tags": [
          "test"
        ],
        "description": "result demo",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "An array of users"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "get all user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully query"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "create a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "args",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully create"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "update a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "args",
            "in": "body",
            "required": true,
            "schema": {
              "properties": {
                "name": {
                  "type": "string"
                },
                "realName": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully update"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "get user by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "user id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully query"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "delete user by if",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "user id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully delete"
          }
        }
      }
    },
    "/changePassword": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "update a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "args",
            "in": "body",
            "required": true,
            "schema": {
              "properties": {
                "username": {
                  "type": "string"
                },
                "byAdmin": {
                  "type": "boolean"
                },
                "oldPassword": {
                  "type": "string"
                },
                "newPassword": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully update"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "user login",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "args",
            "in": "body",
            "required": true,
            "schema": {
              "properties": {
                "name": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully login"
          }
        }
      }
    },
    "/user/logout/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "user logout",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "user id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logout"
          }
        }
      }
    },
    "/userInfo": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "user logout",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "description": "user token",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logout"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "mobile": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}